updated individual signup pages
