+++ src/app/individual-signup/step1/page.tsx
@@ -17,7 +17,6 @@ import AvatarModal from "@/components/modals/chooseAvatarModal";
 import InputComponent from "@/components/inputComponent";
 import LabelGeneratorAndDisplayComp from "@/components/labelGenAndDisplayComponent";
 import InputComponentWithLabelOptions from "@/components/inputComponentWithLabelOptions";
-import ButtonOptionsComponent from "@/components/buttonOptionsComponent";
 import AddHandler from "@/components/addHandler";
 import DeleteHandler from "@/components/deleteHandler";
 
@@ -31,8 +30,6 @@ import {
 type CurrentSchemeType = ButtonColorOption;
 
 const fellowSchema = z.object({
-  name: z.string().min(2, { message: "Your Name is Required" }),
-  email: z.string().email({ message: "Your Email is Required" }),
   smallBio: z
     .string()
     .min(5, { message: "Your Small Bio Must Be More Than 5 Letters" }),
@@ -40,9 +37,6 @@ const fellowSchema = z.object({
   location: z
     .string()
     .min(3, { message: "Your Specific Location is Required" }),
-  locationOptions: z
-    .array(z.string())
-    .min(1, { message: "You Must Have At Least 1 Location Type Selected" }),
   skills: z
     .array(z.string())
     .min(1, { message: "You Must Have At Least 1 Skill Listed" }),
@@ -53,7 +47,6 @@ const fellowSchema = z.object({
     .array(z.string())
     .min(1, { message: "You Must Have At Least 1 Job Title Listed" }),
 });
-// I set the necessary number of skills and jobTitles to 1, but it gives me issues when I only add one item?
 
 type FormData = z.infer<typeof fellowSchema>;
 
@@ -66,15 +59,14 @@ export default function IndividualSignupPage1() {
   const [skills, setSkills] = useState<string[]>([]);
   const [jobTitles, setJobTitles] = useState<string[]>([]);
   const [languages, setLanguages] = useState<string[]>([]);
-  const [locationOptions, setLocationOptions] = useState<string[]>([]);
   const [colorArray, setColorArray] = useState<CurrentSchemeType[]>([]);
   const [secondaryColorArray, setSecondaryColorArray] = useState<
     CurrentSchemeType[]
   >([]);
   const [avatarOptions, setAvatarOptions] = useState({
-    url: fellow?.avatar,
-    shadow: fellow?.shadow,
-    colorScheme: fellow?.colorScheme,
+    url: fellow?.avatar || "/avatars/question.svg",
+    shadow: fellow?.shadow || "drop-shadow-olive",
+    colorScheme: fellow?.colorScheme || "b1",
   });
 
   const {
@@ -89,7 +81,7 @@ export default function IndividualSignupPage1() {
       skills: skills,
       jobTitles: jobTitles,
       country: fellow?.country,
-      locationOptions: locationOptions,
+      languages: languages,
     },
   });
 
@@ -98,34 +90,30 @@ export default function IndividualSignupPage1() {
     console.log(data);
     setFellow({
       ...fellow,
-      name: data.name,
-      email: data.email,
       smallBio: data.smallBio,
       country: data.country,
       location: data.location,
-      locationOptions: locationOptions,
       skills: skills,
       jobTitles: jobTitles,
       avatar: avatarOptions.url,
       shadow: avatarOptions.shadow,
       colorScheme: avatarOptions.colorScheme,
+      languages: languages,
     });
     router.push("/individual-signup/step2");
   };
 
   // handlers for adding, updating, and deleting information tied to States
   const handleAdd = (
-    type: "skills" | "jobTitles" | "country" | "locationOptions" | "languages",
+    type: "skills" | "jobTitles" | "country" | "languages",
     item: any,
   ) => {
-    console.log("handle add - ing", type);
     AddHandler({
       item,
       type,
       setFunctions: {
         skills: setSkills,
         jobTitles: setJobTitles,
-        locationOptions: setLocationOptions,
         languages: setLanguages,
       },
       setValue,
@@ -134,17 +122,15 @@ export default function IndividualSignupPage1() {
   };
 
   const handleDelete = (
-    type: "skills" | "jobTitles" | "languages" | "locationOptions",
+    type: "skills" | "jobTitles" | "languages",
     item: any,
   ) => {
-    console.log("deleting");
     DeleteHandler({
       item,
       type,
       setFunctions: {
         skills: setSkills,
         jobTitles: setJobTitles,
-        locationOptions: setLocationOptions,
         languages: setLanguages,
       },
     });
@@ -162,45 +148,25 @@ export default function IndividualSignupPage1() {
   useEffect(() => {
     setSkills(Array.isArray(fellow?.skills) ? fellow.skills : []);
     setJobTitles(Array.isArray(fellow?.jobTitles) ? fellow.jobTitles : []);
-    setLocationOptions(
-      Array.isArray(fellow?.locationOptions) ? fellow.locationOptions : [],
-    );
+    setLanguages(Array.isArray(fellow?.languages) ? fellow.languages : []);
 
     // Update default values for the form
     setValue("skills", fellow?.skills || []);
     setValue("jobTitles", fellow?.jobTitles || []);
-    setValue("locationOptions", fellow?.locationOptions || []);
+    setValue("languages", fellow?.languages || []);
   }, [fellow, setValue]);
 
   return (
     <div className="IndividualSignupPage flex w-[95vw] max-w-[1600px] flex-grow flex-col items-center justify-center gap-8 self-center pt-6 md:pb-8 md:pt-8">
       <div className="IndividualSignupContainer flex w-[84%] max-w-[1600px] justify-center gap-10 sm:gap-8 md:w-[75%]">
-        <div className="IndividualSignupLeft flex w-[35vw] flex-col">
+        <div className="IndividualSignupLeft mb-8 flex w-[35vw] flex-col">
           <form
             className="IndividualSignupForm xs:pt-8 flex flex-col gap-8"
             onSubmit={handleSubmit(onSubmit)}
           >
-            {/* name input */}
-            <InputComponent
-              type="text"
-              placeholderText="First & Last Name"
-              errors={errors.name}
-              register={register}
-              registerValue="name"
-              defaultValue={fellow?.name}
-              required
-            />
-
-            {/* email input */}
-            <InputComponent
-              type="email"
-              placeholderText="Fantasticemail@emailexample.com"
-              errors={errors.email}
-              register={register}
-              registerValue="email"
-              defaultValue={fellow?.email}
-              required
-            />
+            <h1 className="FellowName -mb-4 ml-8">
+              {fellow?.name || "Test Name"}
+            </h1>
 
             {/* smallBio input */}
             <InputComponent
@@ -238,24 +204,28 @@ export default function IndividualSignupPage1() {
               required
             />
 
-            <ButtonOptionsComponent
-              type="locationOptions"
-              title="location options:"
-              buttons={["remote", "on-site", "hybrid"]}
-              selectedArray={locationOptions}
+            {/* language input & generator */}
+            <LabelGeneratorAndDisplayComp
               handleAdd={handleAdd}
+              errors={errors.languages}
+              selectedArray={languages}
               handleDelete={handleDelete}
-              errors={errors.locationOptions}
+              placeholder="Your Spoken Language(s)"
+              colorArray={colorArray}
+              name="languages"
+              variant="functional"
+              options
+              searchData={languageOptions}
               required
             />
           </form>
         </div>
         <div className="IndividualSignupRight flex w-[35vw] flex-col">
           <Image
-            className={`AvatarImage -mt-14 justify-end self-end ${avatarOptions.shadow}`}
+            className={`AvatarImage -mt-8 justify-end self-end ${avatarOptions.shadow}`}
             src={avatarOptions.url}
-            width={75}
-            height={75}
+            width={65}
+            height={65}
             alt="avatar"
             onClick={() =>
               showModal(<AvatarModal setAvatarOptions={setAvatarOptions} />)
@@ -271,21 +241,6 @@ export default function IndividualSignupPage1() {
           </button>
 
           <div className="SkillsAndJobTitlesContainer flex flex-col gap-8">
-            {/* skills input & generator */}
-            <LabelGeneratorAndDisplayComp
-              handleAdd={handleAdd}
-              errors={errors.skills}
-              selectedArray={skills}
-              handleDelete={handleDelete}
-              placeholder="Your Skills"
-              colorArray={colorArray}
-              name="skills"
-              variant="functional"
-              options
-              searchData={skillsList}
-              required
-            />
-
             {/* job titles generator */}
             <LabelGeneratorAndDisplayComp
               handleAdd={handleAdd}
@@ -299,18 +254,18 @@ export default function IndividualSignupPage1() {
               required
             />
 
-            {/* language input & generator */}
+            {/* skills input & generator */}
             <LabelGeneratorAndDisplayComp
               handleAdd={handleAdd}
-              errors={errors.languages}
-              selectedArray={languages}
+              errors={errors.skills}
+              selectedArray={skills}
               handleDelete={handleDelete}
-              placeholder="Your Spoken Language(s)"
+              placeholder="Your Skills"
               colorArray={colorArray}
-              name="languages"
+              name="skills"
               variant="functional"
               options
-              searchData={languageOptions}
+              searchData={skillsList}
               required
             />
           </div>
